start to breaking up generator
